Enable msvc signed/unsigned mismatch warnings and fix remaining cases

+++ BBGE/Texture.cpp
@@ -449,7 +449,7 @@ ImageTGA *Texture::TGAloadMem(void *mem, int size)
 	ByteBuffer bb(mem, size, ByteBuffer::REUSE);
 
 	ImageTGA *pImageData = NULL;		// This stores our important image data
-	WORD width = 0, height = 0;			// The dimensions of the image
+	uint16_t width = 0, height = 0;		// The dimensions of the image
 	byte length = 0;					// The length in bytes to the pixels
 	byte imageType = 0;					// The image type (RLE, RGB, Alpha...)
 	byte bits = 0;						// The bits per pixel for the image (16, 24, 32)
@@ -604,7 +604,7 @@ ImageTGA *Texture::TGAloadMem(void *mem, int size)
 		byte *pColors = new byte [channels];
 
 		// Load in all the pixel data
-		while(i < width*height)
+		while(i < size_t(width)*height)
 		{
 			// Read in the current color count + 1
 			bb >> rleID;
